eurodist {datasets}	R Documentation

Distances Between European Cities and Between US Cities

Description

The eurodist gives the road distances (in km) between 21 cities in Europe. The data are taken from a table in The Cambridge Encyclopaedia.

UScitiesD gives “straight line” distances between 10 cities in the US.

Usage


eurodist
UScitiesD

Format

dist objects based on 21 and 10 objects, respectively. (You must have the stats package loaded to have the methods for this kind of object available).

Source

Crystal, D. Ed. (1990) The Cambridge Encyclopaedia. Cambridge: Cambridge University Press,

The US cities distances were provided by Pierre Legendre.

[Package datasets version 3.6.1 Index]