sunspot.year {datasets}	R Documentation

Yearly Sunspot Data, 1700-1988

Description

Yearly numbers of sunspots.

Usage

sunspot.year

Format

The univariate time series sunspot.year contains 289 observations, and is of class "ts".

Note

Prior to R 2.0.0 sunspot.year and sunspot.year were copied to the user's workspace (or elsewhere) by data(sunspot). However, as package lattice has a dataset sunspot which can be retrieved by the same command, that usage has been removed.

Source

H. Tong (1996) Non-Linear Time Series. Clarendon Press, Oxford, p. 471.

[Package datasets version 2.1.0 Index]