Indometh {datasets}	R Documentation

Pharmacokinetics of Indomethicin

Description

The Indometh data frame has 66 rows and 3 columns of data on the pharmacokinetics of indomethicin.

Usage

Indometh

Format

This data frame contains the following columns:

Subject

an ordered factor with containing the subject codes. The ordering is according to increasing maximum response.

time

a numeric vector of times at which blood samples were drawn (hr).

conc

a numeric vector of plasma concentrations of indomethicin (mcg/ml).

Details

Each of the six subjects were given an intravenous injection of indomethicin.

Source

Kwan, Breault, Umbenhauer, McMahon and Duggan (1976), Kinetics of Indomethicin absorption, elimination, and enterohepatic circulation in man. Journal of Pharmacokinetics and Biopharmaceutics, 4, 255–280.

Davidian, M. and Giltinan, D. M. (1995) Nonlinear Models for Repeated Measurement Data, Chapman & Hall (section 5.2.4, p. 134)

Pinheiro, J. C. and Bates, D. M. (2000) Mixed-effects Models in S and S-PLUS, Springer.

Examples

require(stats)
fm1 <- nls(conc ~ SSbiexp(time, A1, lrc1, A2, lrc2),
           data = Indometh, subset = Subject == 1)
summary(fm1)

[Package datasets version 2.9.2 Index]