eurodist {datasets}	R Documentation

Distances Between European Cities

Description

The data give the road distances (in km) between 21 cities in Europe. The data are taken from a table in The Cambridge Encyclopaedia.

Usage

eurodist

Format

A dist object based on 21 objects. (You must have the stats package loaded to have the methods for this kind of object available).

Source

Crystal, D. Ed. (1990) The Cambridge Encyclopaedia. Cambridge: Cambridge University Press,

[Package datasets version 2.9.2 Index]